@@ -3,6 +3,7 @@
   "version": "0.0.0",
   "license": "MIT",
   "private": true,
+  "type": "module",
   "resolutions": {
     "camera-controls": "2.9.0"
   },
@@ -10,11 +11,6 @@
     "overrides": {
       "react": "$react",
       "react-dom": "$react-dom"
-    }
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged"
     }
   },
   "lint-staged": {
@@ -31,7 +27,7 @@
     "stub": "pnpm --filter @react-three/fiber stub && pnpm --filter @react-three/eslint-plugin stub",
     "dev": "pnpm stub",
     "ci": "pnpm build && pnpm typecheck && pnpm eslint && pnpm dev && pnpm test && pnpm format",
-    "prepare": "husky install",
+    "prepare": "husky",
     "eslint": "eslint packages/**/src/**/*.{ts,tsx}",
     "eslint:fix": "pnpm run eslint --fix",
     "test": "vitest run --coverage",
@@ -47,49 +43,46 @@
     "verify-bundles": "node scripts/verify-bundles.js"
   },
   "devDependencies": {
-    "@babel/core": "^7.22.0",
-    "@babel/preset-env": "7.16.11",
-    "@babel/preset-react": "7.16.7",
-    "@babel/preset-typescript": "^7.16.7",
-    "@changesets/changelog-git": "^0.1.11",
-    "@changesets/cli": "^2.22.0",
-    "@rollup/plugin-alias": "^5.1.1",
-    "@testing-library/dom": "^10.0.0",
-    "@testing-library/react": "^16.0.0",
+    "@babel/core": "^7.28.6",
+    "@babel/preset-env": "^7.28.6",
+    "@babel/preset-react": "^7.28.5",
+    "@babel/preset-typescript": "^7.28.5",
+    "@changesets/changelog-git": "^0.2.1",
+    "@changesets/cli": "^2.29.8",
+    "@rollup/plugin-alias": "^6.0.0",
+    "@testing-library/dom": "^10.4.1",
+    "@testing-library/react": "^16.3.1",
     "@types/node": "^22.0.0",
     "@types/offscreencanvas": "^2019.7.3",
-    "@types/prettier": "^2.7.3",
-    "@types/react": "^19.0.1",
-    "@types/react-dom": "^19.0.1",
-    "@types/react-reconciler": "^0.32.0",
-    "@types/scheduler": "0.23.0",
+    "@types/react": "^19.2.8",
+    "@types/react-dom": "^19.2.3",
+    "@types/react-reconciler": "^0.32.3",
+    "@types/scheduler": "^0.26.0",
     "@types/three": "^0.181.0",
-    "@typescript-eslint/eslint-plugin": "^5.17.0",
-    "@typescript-eslint/parser": "^5.17.0",
-    "@vitest/coverage-v8": "^4.0.16",
-    "babel-plugin-module-resolver": "^5.0.0",
-    "eslint": "^8.12.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-import-resolver-alias": "^1.1.2",
-    "eslint-plugin-import": "^2.25.4",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.29.4",
-    "eslint-plugin-react-hooks": "^4.4.0",
-    "eslint-plugin-vitest": "^0.3.26",
-    "husky": "^7.0.4",
+    "@vitest/coverage-v8": "^4.0.17",
+    "babel-plugin-module-resolver": "^5.0.2",
+    "eslint": "^9.39.2",
+    "eslint-config-prettier": "^10.1.8",
+    "eslint-plugin-prettier": "^5.5.5",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-vitest": "^0.5.4",
+    "globals": "^16.4.0",
+    "typescript-eslint": "^8.53.0",
+    "husky": "^9.1.7",
     "jsdom": "^27.4.0",
-    "lint-staged": "^12.3.7",
-    "prettier": "^2.6.1",
-    "pretty-quick": "^3.1.3",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
+    "lint-staged": "^16.2.7",
+    "prettier": "^3.8.0",
+    "pretty-quick": "^4.2.2",
+    "react": "^19.2.3",
+    "react-dom": "^19.2.3",
     "react-nil": "^2.0.0",
     "react-reconciler": "^0.33.0",
     "three": "^0.181.0",
     "typescript": "^5.9.3",
     "unbuild": "^3.6.1",
-    "vite": "^6.0.0",
-    "vitest": "^4.0.16"
+    "vite": "^7.3.1",
+    "vitest": "^4.0.17"
   },
   "packageManager": "pnpm@9.15.2"
 }
