@@ -19,7 +19,7 @@ import { isOrthographicCamera } from './is'
 export function calculateDpr(dpr: Dpr): number {
   // Err on the side of progress by assuming 2x dpr if we can't detect it
   // This will happen in workers where window is defined but dpr isn't.
-  const target = typeof window !== 'undefined' ? window.devicePixelRatio ?? 2 : 1
+  const target = typeof window !== 'undefined' ? (window.devicePixelRatio ?? 2) : 1
   return Array.isArray(dpr) ? Math.min(Math.max(dpr[0], target), dpr[1]) : dpr
 }
 
