@@ -14,8 +14,8 @@ export type MathTypes = MathRepresentation | THREE.Euler | THREE.Color
 export type MathType<T extends MathTypes> = T extends THREE.Color
   ? Args<typeof THREE.Color> | THREE.ColorRepresentation
   : T extends VectorRepresentation | THREE.Layers | THREE.Euler
-  ? T | MutableOrReadonlyParameters<T['set']> | number
-  : T | MutableOrReadonlyParameters<T['set']>
+    ? T | MutableOrReadonlyParameters<T['set']> | number
+    : T | MutableOrReadonlyParameters<T['set']>
 
 export type MathProps<P> = {
   [K in keyof P as P[K] extends MathTypes ? K : never]: P[K] extends MathTypes ? MathType<P[K]> : never
